fix: reorganize and rename folder structure
