+++ src/components/index.ts
@@ -1,9 +1,5 @@
-export * from './empty-state/empty-state';
 export * from './error-boundary/error-boundary';
 export * from './card/card';
 export * from './footer/footer';
 export * from './header/header';
-export * from './pagination/pagination';
-export * from './search/search';
-export * from './spinner/spinner';
-export * from './ui';
+export * from './shared';
