+++ src/components/shared/search/search.css
