+++ src/components/shared/spinner/spinner.css
